requirements.txt
# Core
anthropic>=0.40.0
cryptography>=42.0.0
openai>=1.50.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Prediction / ML
xgboost>=2.0.0
scikit-learn>=1.5.0
numpy>=1.26.0
pandas>=2.0.0

# Data sources
feedparser>=6.0.11
praw>=7.7.0            # Reddit API

# Scheduling
schedule>=1.2.0

# Monitoring
prometheus-client>=0.20.0

# Optional: Google AI (Gemini)
google-generativeai>=0.8.0
